+++ app/(docs)/docs/animatedcard/animatedcard.tsx
