+++ app/registry/animatedcard/animatedcard.tsx
