+++ public/r/animated-card.json
@@ -15,7 +15,7 @@
     },
     {
       "path": "app/registry/animatedcard/animatedcarddemo.tsx",
-      "content": "'use client'\nimport AnimatedCard from './animatedCard';\nimport { easeInOut, motion } from 'framer-motion';\nimport { useState } from 'react';\n\n\nconst tools = [\n  {\n    imgSrc: '/tailwind.svg',\n    title: 'Tailwind CSS',\n\n    aboutProduct: 'A utility-first CSS framework for rapidly building custom UIs.',\n  },\n  {\n    imgSrc: '/nextjs.svg',\n    title: 'Next.js',\n\n    aboutProduct:\n      'A React framework for production with server-side rendering and static site generation.',\n  },\n  {\n    imgSrc: '/shadcn.svg',\n    title: 'Shadcn UI',\n\n    aboutProduct: 'Beautiful, accessible components built using Radix UI and Tailwind CSS.',\n  },\n  {\n    imgSrc: 'aceternity.svg',\n    title: 'Aceternity UI',\n\n    aboutProduct: 'Beautiful, accessible components built using Shadcn UI and Tailwind CSS.',\n  },\n];\n\nexport default function AnimatedCardDemo() {\n  const [touchComponent, setTouchComponent] = useState(false);\n\n  const rotateDegree = [-20, -10, 0, 20];\n  const xAxis = [-300, -50, 240, 330];\n  const yAxis = [-60, -120, -140, -100];\n  const initialRotation = [0, 6, 12, 17];\n  const zIndex = [40, 30, 20, 10];\n\n  return (\n    <div\n      className=\"hidden md:flex justify-center min-h-screen items-center\"\n      onClick={() => setTouchComponent(!touchComponent)}\n    >\n      {tools.map((tool, ind) => (\n        <motion.div\n          key={ind}\n          initial={{ x: 0, y: 0 }}\n          animate={\n            touchComponent\n              ? { x: xAxis[ind], y: yAxis[ind], rotate: rotateDegree[ind] }\n              : { x: 0, y: 0 }\n          }\n          transition={{ ease: easeInOut }}\n          style={{\n            zIndex: zIndex[ind],\n            rotate: initialRotation[ind],\n          }}\n          className=\"absolute flex justify-center items-center \"\n        >\n          <AnimatedCard imgSrc={tool.imgSrc} title={tool.title} aboutProduct={tool.aboutProduct} />\n        </motion.div>\n      ))}\n    </div>\n  );\n}\n",
+      "content": "'use client'\nimport AnimatedCard from './animatedcard';\nimport { easeInOut, motion } from 'framer-motion';\nimport { useState } from 'react';\n\n\nconst tools = [\n  {\n    imgSrc: '/tailwind.svg',\n    title: 'Tailwind CSS',\n\n    aboutProduct: 'A utility-first CSS framework for rapidly building custom UIs.',\n  },\n  {\n    imgSrc: '/nextjs.svg',\n    title: 'Next.js',\n\n    aboutProduct:\n      'A React framework for production with server-side rendering and static site generation.',\n  },\n  {\n    imgSrc: '/shadcn.svg',\n    title: 'Shadcn UI',\n\n    aboutProduct: 'Beautiful, accessible components built using Radix UI and Tailwind CSS.',\n  },\n  {\n    imgSrc: 'aceternity.svg',\n    title: 'Aceternity UI',\n\n    aboutProduct: 'Beautiful, accessible components built using Shadcn UI and Tailwind CSS.',\n  },\n];\n\nexport default function AnimatedCardDemo() {\n  const [touchComponent, setTouchComponent] = useState(false);\n\n  const rotateDegree = [-20, -10, 0, 20];\n  const xAxis = [-300, -50, 240, 330];\n  const yAxis = [-60, -120, -140, -100];\n  const initialRotation = [0, 6, 12, 17];\n  const zIndex = [40, 30, 20, 10];\n\n  return (\n    <div\n      className=\"hidden md:flex justify-center min-h-screen items-center\"\n      onClick={() => setTouchComponent(!touchComponent)}\n    >\n      {tools.map((tool, ind) => (\n        <motion.div\n          key={ind}\n          initial={{ x: 0, y: 0 }}\n          animate={\n            touchComponent\n              ? { x: xAxis[ind], y: yAxis[ind], rotate: rotateDegree[ind] }\n              : { x: 0, y: 0 }\n          }\n          transition={{ ease: easeInOut }}\n          style={{\n            zIndex: zIndex[ind],\n            rotate: initialRotation[ind],\n          }}\n          className=\"absolute flex justify-center items-center \"\n        >\n          <AnimatedCard imgSrc={tool.imgSrc} title={tool.title} aboutProduct={tool.aboutProduct} />\n        </motion.div>\n      ))}\n    </div>\n  );\n}\n",
       "type": "registry:component",
       "target": "components/spectrumui/animatedcarddemo.tsx"
     }
